[FIX] apply algorithm on unlink method only in intercompany_trade case;

+++ intercompany_trade_account/model/account_invoice_line.py
@@ -212,18 +212,19 @@ class AccountInvoiceLine(Model):
             ctx['intercompany_trade_do_not_propagate'] = True
             for ail in self.browse(
                     cr, uid, ids, context=context):
-                ai = ail.invoice_id
-                rit = ai_obj._get_intercompany_trade_config(
-                    cr, uid, ai.partner_id.id, ai.company_id.id, ai.type,
-                    context=context)
-                if ail.invoice_id.type in ('in_invoice', 'in_refund'):
-                    other_uid = rit.supplier_user_id.id
-                else:
-                    other_uid = rit.customer_user_id.id
-                self.unlink(
-                    cr, other_uid,
-                    [ail.intercompany_trade_account_invoice_line_id.id],
-                    context=ctx)
+                if ail.intercompany_trade:
+                    rit = ai_obj._get_intercompany_trade_config(
+                        cr, uid, ail.invoice_id.partner_id.id,
+                        ail.invoice_id.company_id.id, ail.invoice_id.type,
+                        context=context)
+                    if ail.invoice_id.type in ('in_invoice', 'in_refund'):
+                        other_uid = rit.supplier_user_id.id
+                    else:
+                        other_uid = rit.customer_user_id.id
+                    self.unlink(
+                        cr, other_uid,
+                        [ail.intercompany_trade_account_invoice_line_id.id],
+                        context=ctx)
         res = super(AccountInvoiceLine, self).unlink(
             cr, uid, ids, context=context)
         return res
